Back to original mail until implemented self sending mail
